Second commit financial calculator

+++ python_introduction/finance_calculator.py
@@ -4,7 +4,7 @@ monthlyExpenses = int(input("Enter your total monthly expenses: "))
 
 monthlySavings = monthlyIncome - monthlyExpenses
 
-projectedSavings = (monthlySavings * 12 + (monthlySavings * 12 * 0.05))
+projectedSavings = int((monthlySavings * 12 + (monthlySavings * 12 * 0.05)))
 
 print(f"Your monthly savings are ${monthlySavings}.")
 print(f"Projected savings after one year, with interest, is: ${projectedSavings}.")
